Add a --context-depth option to vg map
Now vg map will collect edge-connected nodes out to a certain depth when
building its subgraph for a thread for aligning. The defaukt depth is 1.
This is important for when you are aligning to graphs which do not have
IDs in a nice linear order along the graph (or which can't due to the
graph looping around a lot).

This is accomplished with index->expand_context(), which is potentially
inefficient (due to scanning the edges over and over) for large context
values.