Revise path rank handling
Now instead of path_mapping_order storing an order number for each
mapping, we just use the rank field on the mapping. We also try and set
reasonable ranks when mappings are appended, and don't allow duplicate
rank values to slip past when recalculating the ranks based on the
actual list ordering.

We also add a mappings_by_rank map that can get the mapping with a given
path name and rank efficiently.

This eliminates the O(n^2) factor when adding many mappings to the same
node to a path, as we only have to check if a mapping with the same rank
has already been added (by an efficient lookup in mappings_by_rank),
rather than having to scan the set of all mappings to that node for one
that matches in rank and orientation.

The Paths::has_mapping method has been changed so you can no longer pass
in a Mapping&. It may need more overloads, but the semantics of querying
by a Mapping reference weren't entirely clear (for example, the actual
edits were never checked), so I've decided to switch over to overloads
that query on particular combinations of Mapping properties instead.

This commit includes some fixes to make it work correctly with the MSGA
path inclusion, which the original commit on the bake-off branch did not
have.